chore: 학습 상세 페이지 컴포넌트에서 스타일 분리
